refactor: move manager.py to morphodict.paradigm package
