+++ src/morphodict/paradigm/manager.py
